+++ coxeter/shapes/circle.py
@@ -166,7 +166,8 @@ class Circle(Shape2D):
             points (:math:`(N, 3)` :class:`numpy.ndarray`):
                 The points to test.
 
-        Returns:
+        Returns
+        -------
             :math:`(N, )` :class:`numpy.ndarray`:
                 Boolean array indicating which points are contained in the
                 circle.