+++ coxeter/shapes/ellipsoid.py
@@ -184,7 +184,8 @@ class Ellipsoid(Shape3D):
             points (:math:`(N, 3)` :class:`numpy.ndarray`):
                 The points to test.
 
-        Returns:
+        Returns
+        -------
             :math:`(N, )` :class:`numpy.ndarray`:
                 Boolean array indicating which points are contained in the
                 ellipsoid.