+++ coxeter/shapes/sphere.py
@@ -131,7 +131,8 @@ class Sphere(Shape3D):
             points (:math:`(N, 3)` :class:`numpy.ndarray`):
                 The points to test.
 
-        Returns:
+        Returns
+        -------
             :math:`(N, )` :class:`numpy.ndarray`:
                 Boolean array indicating which points are contained in the
                 sphere.